+++ R/fem.R
@@ -398,11 +398,24 @@ fm_fem.fm_mesh_3d <- function(mesh, order = 2, ...) {
 #' @description `r lifecycle::badge("experimental")`
 #'   Compute effective sizes of faces/cells and vertices in a mesh
 #' @param ... Passed on to submethods
-#' @returns A `list` with elements `face` and `vertex` for 2D meshes, or `cell`
-#'   and `vertex` for 3D meshes. The elements are vectors of effective sizes of
-#'   the faces/cells and vertices, respectively. For 2D meshes, also
-#'   `face_edge`, a matrix with one row per triangle and 3 columns, with edge
-#'   lengths for the edge opposing each triangle vertex.
+#' @returns A `list` with elements of simplex size information. For 2D meshes:
+#'   \describe{
+#'   \item{`face`}{Vector with the area of each triangle}
+#'   \item{`vertex`}{Vector with the triangle area apportioned to each vertex}
+#'   \item{`face_edge`}{A matrix with one row per triangle and 3 columns, with
+#'     edge lengths for the edge opposing each triangle vertex.}
+#'   }
+#'   For 3D meshes:
+#'   \describe{
+#'   \item{`cell`}{Vector with the volume of each tetrahedron}
+#'   \item{`vertex`}{Vector with the tetrahedron volume apportioned to each
+#'   vertex}
+#'   \item{`cell_face`}{A matrix with one row per cell and 4 columns, with
+#'   triangle areas for the triangle opposing each tetrahedron vertex.}
+#'   \item{`cell_edge`}{A matrix with one row per cell and 4 columns, with
+#'     edge lengths for the edge anchored at each vertex, pointing to the next
+#'     vertex in the internal ordering.}
+#'   }
 #' @export
 #' @examples
 #' str(fm_sizes(fmexample$mesh))
@@ -424,9 +437,10 @@ fm_sizes.fm_mesh_2d <- function(mesh, ...) {
   e1 <- v3 - v2
   e2 <- v1 - v3
   e3 <- v2 - v1
-  areas_t <- rowSums((row_cross_product(e1, e2) +
-                        row_cross_product(e2, e3) +
-                        row_cross_product(e3, e1))^2)^0.5 / 6
+  areas_t <- rowSums(
+    (row_cross_product(e1, e2) +
+    row_cross_product(e2, e3) +
+    row_cross_product(e3, e1))^2)^0.5 / 6
   lengths_t <- cbind(
     rowSums(e1 * e1)^0.5,
     rowSums(e2 * e2)^0.5,
@@ -456,6 +470,18 @@ fm_sizes.fm_mesh_3d <- function(mesh, ...) {
   e3 <- v4 - v3
   e4 <- v1 - v4
   vols_t <- abs(row_volume_product(e1, e2, e3)) / 6
+  cell_face_t <- cbind(
+    rowSums(row_cross_product(e2, e3)^2)^0.5,
+    rowSums(row_cross_product(e3, e4)^2)^0.5,
+    rowSums(row_cross_product(e4, e1)^2)^0.5,
+    rowSums(row_cross_product(e1, e2)^2)^0.5
+  ) / 2
+  cell_edge_t <- cbind(
+    rowSums(e1^2)^0.5,
+    rowSums(e2^2)^0.5,
+    rowSums(e3^2)^0.5,
+    rowSums(e4^2)^0.5
+  )
 
   c0 <- Matrix::sparseMatrix(
     i = as.vector(mesh$graph$tv),
@@ -465,5 +491,5 @@ fm_sizes.fm_mesh_3d <- function(mesh, ...) {
   )
   vols_v <- Matrix::diag(c0)
 
-  list(cell = vols_t, vertex = vols_v)
+  list(cell = vols_t, cell_face = cell_face_t, cell_edge_t, vertex = vols_v)
 }