+++ NEWS.md
@@ -5,6 +5,8 @@
 * Speed up of `fm_int()` for 2D meshes by internally storing the barycentric
   coordinate information during the integration scheme construction
   (version `0.6.1.9002`)
+* Add `face_edge` (for 2D meshes), `cell_face`, and `cell_edge` (for 3D meshes)
+  information to `fm_sizes()` output (version `0.6.1.9004`)
 
 ## Bug fixes